refactor: extract archive api handlers
